refactor(components): new distribution
