+++ .eslintrc.json
@@ -1,3 +1,6 @@
 {
-  "extends": ["next/core-web-vitals","next","prettier"]
+  "extends": ["next/core-web-vitals","next","prettier"],
+  "rules": {
+    "no-console": 1
+  }
 }
